# apps/web/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
      "Allocation": {
        "properties": {
          "actual_arrival_tick": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Arrival Tick"
          },
          "created_tick": {
            "title": "Created Tick",
            "type": "integer"
          },
          "departure_tick": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Departure Tick"
          },
          "destination_station_id": {
            "title": "Destination Station Id",
            "type": "string"
          },
          "expected_arrival_tick": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Arrival Tick"
          },
          "failure_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Reason"
          },
          "fuel_type": {
            "enum": [
              "DIESEL",
              "PETROL",
              "OCTANE"
            ],
            "title": "Fuel Type",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "idempotency_key": {
            "title": "Idempotency Key",
            "type": "string"
          },
          "quantity": {
            "title": "Quantity",
            "type": "number"
          },
          "route_id": {
            "title": "Route Id",
            "type": "string"
          },
          "source_depot_id": {
            "title": "Source Depot Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "PENDING",
              "IN_TRANSIT",
              "ARRIVED",
              "FAILED",
              "CANCELLED"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "idempotency_key",
          "source_depot_id",
          "destination_station_id",
          "route_id",
          "fuel_type",
          "quantity",
          "created_tick",
          "departure_tick",
          "expected_arrival_tick",
          "actual_arrival_tick",
          "status",
          "failure_reason"
        ],
        "title": "Allocation",
        "type": "object"
      },
...
      "AutomationState": {
        "properties": {
          "kill_switch": {
            "description": "When true, nothing executes, in any mode.",
            "title": "Kill Switch",
            "type": "boolean"
          },
          "mode": {
            "enum": [
              "ADVISORY",
              "GUARDED_AUTO",
              "MANUAL_DEMO"
            ],
            "title": "Mode",
            "type": "string"
          }
        },
        "required": [
          "mode",
          "kill_switch"
        ],
        "title": "AutomationState",
        "type": "object"
      },
...
      "DemandObservation": {
        "properties": {
          "demand_liters": {
            "title": "Demand Liters",
            "type": "number"
          },
          "fuel_type": {
            "enum": [
              "DIESEL",
              "PETROL",
              "OCTANE"
            ],
            "title": "Fuel Type",
            "type": "string"
          },
          "served_liters": {
            "title": "Served Liters",
            "type": "number"
          },
          "station_id": {
            "title": "Station Id",
            "type": "string"
          },
          "tick": {
            "title": "Tick",
            "type": "integer"
          },
          "unmet_liters": {
            "title": "Unmet Liters",
            "type": "number"
          }
        },
        "required": [
          "station_id",
          "fuel_type",
          "tick",
          "demand_liters",
          "served_liters",
          "unmet_liters"
        ],
        "title": "DemandObservation",
        "type": "object"
      },
      "Depot": {
        "properties": {
          "capacity": {
            "additionalProperties": {
              "type": "number"
            },
            "propertyNames": {
              "enum": [
                "DIESEL",
                "PETROL",
                "OCTANE"
              ]
            },
            "title": "Capacity",
            "type": "object"
          },
          "dispatch_capacity_per_tick": {
            "title": "Dispatch Capacity Per Tick",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "inventory": {
            "additionalProperties": {
              "type": "number"
            },
            "propertyNames": {
              "enum": [
                "DIESEL",
                "PETROL",
                "OCTANE"
              ]
            },
            "title": "Inventory",
            "type": "object"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "region_id": {
            "title": "Region Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "OPEN",
              "CONSTRAINED"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "region_id",
          "status",
          "dispatch_capacity_per_tick",
          "capacity",
          "inventory"
        ],
        "title": "Depot",
        "type": "object"
      },
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "Forecast": {
        "description": "Forecast of demand_liters (not served_liters) per tick.",
        "properties": {
          "calibration": {
            "description": "alpha in forecast = alpha x structural demand.",
            "title": "Calibration",
            "type": "number"
          },
          "fuel_type": {
            "enum": [
              "DIESEL",
              "PETROL",
              "OCTANE"
            ],
            "title": "Fuel Type",
            "type": "string"
          },
          "liters_per_tick": {
            "items": {
              "type": "number"
            },
            "title": "Liters Per Tick",
            "type": "array"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "start_tick": {
            "title": "Start Tick",
            "type": "integer"
          },
          "station_id": {
            "title": "Station Id",
            "type": "string"
          }
        },
        "required": [
          "station_id",
          "fuel_type",
          "start_tick",
          "liters_per_tick",
          "calibration",
          "method"
        ],
        "title": "Forecast",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InventoryProjection": {
        "properties": {
          "fuel_type": {
            "enum": [
              "DIESEL",
              "PETROL",
              "OCTANE"
            ],
            "title": "Fuel Type",
            "type": "string"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/ProjectionPoint"
            },
            "title": "Points",
            "type": "array"
          },
          "safety_stock": {
            "title": "Safety Stock",
            "type": "number"
          },
          "station_id": {
            "title": "Station Id",
            "type": "string"
          }
        },
        "required": [
          "station_id",
          "fuel_type",
          "safety_stock",
          "points"
        ],
        "title": "InventoryProjection",
        "type": "object"
      },
      "NetworkState": {
        "description": "One consistent view of the world. Built only by the API; intelligence never fetches.\n\n`demand_history` holds recent observations, oldest first.",
        "properties": {
          "allocations": {
            "items": {
              "$ref": "#/components/schemas/Allocation"
            },
            "title": "Allocations",
            "type": "array"
          },
          "demand_history": {
            "items": {
              "$ref": "#/components/schemas/DemandObservation"
            },
            "title": "Demand History",
            "type": "array"
          },
          "depots": {
            "items": {
              "$ref": "#/components/schemas/Depot"
            },
            "title": "Depots",
            "type": "array"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/SimEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/SnapshotMeta"
          },
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SimMetrics"
              },
              {
                "type": "null"
              }
            ]
          },
          "regions": {
            "items": {
              "$ref": "#/components/schemas/Region"
            },
            "title": "Regions",
            "type": "array"
          },
          "routes": {
            "items": {
              "$ref": "#/components/schemas/Route"
            },
            "title": "Routes",
            "type": "array"
          },
          "run": {
            "$ref": "#/components/schemas/RunInfo"
          },
          "stations": {
            "items": {
              "$ref": "#/components/schemas/Station"
            },
            "title": "Stations",
            "type": "array"
          },
          "supply_arrivals": {
            "items": {
              "$ref": "#/components/schemas/SupplyArrival"
            },
            "title": "Supply Arrivals",
            "type": "array"
          }
        },
        "required": [
          "meta",
          "run",
          "regions",
          "depots",
          "stations",
          "routes",
          "supply_arrivals",
          "events",
          "allocations",
          "demand_history"
        ],
        "title": "NetworkState",
        "type": "object"
      },
      "Plan": {
        "properties": {
          "forecasts": {
            "items": {
              "$ref": "#/components/schemas/Forecast"
            },
            "title": "Forecasts",
            "type": "array"
          },
          "planner_version": {
            "description": "Planner that produced the recommendations, e.g. 'order-up-to-v0' or 'fallback'.",
            "title": "Planner Version",
            "type": "string"
          },
          "projections": {
            "items": {
              "$ref": "#/components/schemas/InventoryProjection"
            },
            "title": "Projections",
            "type": "array"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/Recommendation"
            },
            "title": "Recommendations",
            "type": "array"
          },
          "risks": {
            "items": {
              "$ref": "#/components/schemas/RiskAssessment"
            },
            "title": "Risks",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "tick": {
            "title": "Tick",
            "type": "integer"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "tick",
          "run_id",
          "planner_version",
          "forecasts",
          "projections",
          "risks",
          "recommendations"
        ],
        "title": "Plan",
        "type": "object"
      },
...
      "Recommendation": {
        "properties": {
          "alternatives": {
            "items": {
              "$ref": "#/components/schemas/AlternativeAction"
            },
            "title": "Alternatives",
            "type": "array"
          },
          "current_inventory": {
            "title": "Current Inventory",
            "type": "number"
          },
          "dispatch_tick": {
            "title": "Dispatch Tick",
            "type": "integer"
          },
          "expected_arrival_tick": {
            "title": "Expected Arrival Tick",
            "type": "integer"
          },
          "expiry_tick": {
            "description": "Not executable after this tick; the orchestrator must replan.",
            "title": "Expiry Tick",
            "type": "integer"
          },
          "factors": {
            "items": {
              "type": "string"
            },
            "title": "Factors",
            "type": "array"
          },
          "fuel_type": {
            "enum": [
              "DIESEL",
              "PETROL",
              "OCTANE"
            ],
            "title": "Fuel Type",
            "type": "string"
          },
          "generated_tick": {
            "title": "Generated Tick",
            "type": "integer"
          },
          "id": {
            "description": "Deterministic: same NetworkState + Policy -> same id.",
            "title": "Id",
            "type": "string"
          },
          "planner_version": {
            "title": "Planner Version",
            "type": "string"
          },
          "priority": {
            "enum": [
              "OK",
              "WATCH",
              "HIGH",
              "CRITICAL"
            ],
            "title": "Priority",
            "type": "string"
          },
          "projected_inventory_at_arrival": {
            "title": "Projected Inventory At Arrival",
            "type": "number"
          },
          "projected_safety_breach_tick": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projected Safety Breach Tick"
          },
          "projected_stockout_tick": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projected Stockout Tick"
          },
          "reason_codes": {
            "items": {
              "enum": [
                "SAFETY_STOCK_BREACH",
                "STOCKOUT_BEFORE_ARRIVAL",
                "CONNECTIVITY_RISK",
                "SINGLE_SOURCE",
                "STATION_OUTAGE",
                "SCARCITY_LIMITED",
                "ROUTE_DISRUPTED"
              ],
              "type": "string"
            },
            "title": "Reason Codes",
            "type": "array"
          },
          "request": {
            "$ref": "#/components/schemas/AllocationRequest"
          },
          "station_id": {
            "title": "Station Id",
            "type": "string"
          },
          "stress_margin_liters": {
            "description": "Projected inventory at arrival minus safety stock, without this shipment.",
            "title": "Stress Margin Liters",
            "type": "number"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "id",
          "station_id",
          "fuel_type",
          "priority",
          "reason_codes",
          "request",
          "generated_tick",
          "expiry_tick",
          "dispatch_tick",
          "expected_arrival_tick",
          "current_inventory",
          "projected_inventory_at_arrival",
          "projected_safety_breach_tick",
          "projected_stockout_tick",
          "stress_margin_liters",
          "summary",
          "factors",
          "alternatives",
          "planner_version"
        ],
        "title": "Recommendation",
        "type": "object"
      },
      "RecommendationState": {
        "properties": {
          "allocation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allocation Id"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "recommendation_id": {
            "title": "Recommendation Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "PROPOSED",
              "APPROVED",
              "EXECUTING",
              "DONE",
              "FAILED",
              "EXPIRED",
              "SUPERSEDED",
              "REJECTED"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_tick": {
            "title": "Updated Tick",
            "type": "integer"
          }
        },
        "required": [
          "recommendation_id",
          "status",
          "updated_tick"
        ],
        "title": "RecommendationState",
        "type": "object"
      },
...
      "RiskAssessment": {
        "properties": {
          "current_inventory": {
            "title": "Current Inventory",
            "type": "number"
          },
          "earliest_arrival_tick": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Earliest tick a new shipment could land; null if unreachable.",
            "title": "Earliest Arrival Tick"
          },
          "fuel_type": {
            "enum": [
              "DIESEL",
              "PETROL",
              "OCTANE"
            ],
            "title": "Fuel Type",
            "type": "string"
          },
          "level": {
            "enum": [
              "OK",
              "WATCH",
              "HIGH",
              "CRITICAL"
            ],
            "title": "Level",
            "type": "string"
          },
          "projected_inventory_at_arrival": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projected Inventory At Arrival"
          },
          "projected_safety_breach_tick": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projected Safety Breach Tick"
          },
          "projected_stockout_tick": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projected Stockout Tick"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "reason_codes": {
            "items": {
              "enum": [
                "SAFETY_STOCK_BREACH",
                "STOCKOUT_BEFORE_ARRIVAL",
                "CONNECTIVITY_RISK",
                "SINGLE_SOURCE",
                "STATION_OUTAGE",
                "SCARCITY_LIMITED",
                "ROUTE_DISRUPTED"
              ],
              "type": "string"
            },
            "title": "Reason Codes",
            "type": "array"
          },
          "redundancy": {
            "description": "Number of currently feasible routes to this station for this fuel.",
            "title": "Redundancy",
            "type": "integer"
          },
          "safety_stock": {
            "title": "Safety Stock",
            "type": "number"
          },
          "station_id": {
            "title": "Station Id",
            "type": "string"
          }
        },
        "required": [
          "station_id",
          "fuel_type",
          "level",
          "reason_codes",
          "current_inventory",
          "safety_stock",
          "projected_inventory_at_arrival",
          "projected_safety_breach_tick",
          "projected_stockout_tick",
          "earliest_arrival_tick",
          "redundancy",
          "reason"
        ],
        "title": "RiskAssessment",
        "type": "object"
      },
...
      "SimControlRequest": {
        "properties": {
          "action": {
            "enum": [
              "run",
              "pause",
              "step"
            ],
            "title": "Action",
            "type": "string"
          }
        },
        "required": [
          "action"
        ],
        "title": "SimControlRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "FuelOps AI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/automation": {
      "get": {
        "operationId": "get_automation_api_automation_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationState"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Automation",
        "tags": [
          "operations"
        ]
      },
      "put": {
        "operationId": "set_automation_api_automation_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutomationState"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Automation",
        "tags": [
          "operations"
        ]
      }
    },
...
    "/api/plan": {
      "get": {
        "operationId": "get_plan_api_plan_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Plan",
        "tags": [
          "operations"
        ]
      }
    },
    "/api/recommendations/{recommendation_id}/approve": {
      "post": {
        "operationId": "approve_api_recommendations__recommendation_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "recommendation_id",
            "required": true,
            "schema": {
              "title": "Recommendation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationState"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Approve",
        "tags": [
          "operations"
        ]
      }
    },
    "/api/recommendations/{recommendation_id}/reject": {
      "post": {
        "operationId": "reject_api_recommendations__recommendation_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "recommendation_id",
            "required": true,
            "schema": {
              "title": "Recommendation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Reject",
        "tags": [
          "operations"
        ]
      }
    },
    "/api/sim/control": {
      "post": {
        "description": "Demo controls proxied to simulator /admin/*. Never used by planning logic.",
        "operationId": "sim_control_api_sim_control_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimControlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Sim Control Api Sim Control Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Implemented"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "summary": "Sim Control",
        "tags": [
          "operations"
        ]
      }
    },
    "/api/state": {
      "get": {
        "operationId": "get_state_api_state_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkState"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get State",
        "tags": [
          "operations"
        ]
      }
    }
  }
}